Use a component index to speed up archetype filtering (#70)
This is for cases with many archetypes. Instead of iterating all archetypes and chacking the mask, we have an index that maps from each component type to all archetypes that contain that type.

When query iteration starts, the query's component with the least archetypes is determined, and it is used to pre-select the archetypes to iterate.